!! broken
I made some additions while thinking about how to deal with evalution of
a token. If a treat a token as (oper head tail), then I should be able
to recursively call head->oper() and tail->oper() applying values to
either the token itself or in place of where it used to be.

this got me thinking about types and how to handle various operations.
mnay languages allow cross type operations and treat

   (* 2 "ndht") -> "ndhtndht"
   would (* 2 join("ndht", ' ')) -> "ndht ndht" or "ndht ndht "??

this means that the multiplication operator should now what its types
are. even deeper than that, I a node to be able to different values.
This got me thinking of templating the node.

   node<integer>
   node<string>

starting over and working to do basic operations. (+ 2 3) and (* 2 "a").