application of function to token->oper
but, I'm not sure if it's actually being called. Still don't have
adequate printing of the token internals yet.

added a check at the end of the parser to prevent incomplete (+ 3)
expressions being "valid".

ended up moving the "operations.hpp" content into "token.hpp". this
should not need to happen, but I have an order of declaratio issue
that needs to be identified to keep them separate.